requirements.txt
# Requirements for Cognitio Pro LMS
# Production Deployment - Complete Package List

# Core Flask Framework
Flask==3.1.2
Flask-SocketIO==5.6.0
werkzeug==3.1.4

# Database
mysql-connector-python==8.0.33

# Security Packages
flask-talisman==1.1.0          # HTTPS enforcement, security headers, CSP
flask-limiter==3.8.0           # Rate limiting
Flask-WTF>=1.2.0               # CSRF protection (compatible with Flask 3.x)
bleach==6.1.0                  # HTML sanitization (XSS prevention)
python-dotenv==1.0.0           # Environment variables for secrets
redis==5.0.1                   # For Flask-Limiter storage (production)
argon2-cffi==23.1.0            # Password hashing
cryptography==42.0.2           # Encryption utilities
pyjwt==2.8.0                   # JWT tokens
python-json-logger==2.0.7      # Structured logging
flask-ipban==1.1.5             # IP-based banning

# Computer Vision & Media Processing
opencv-python==4.8.1.78

# Excel/CSV Processing
openpyxl==3.1.2

# Optional: Production Server
gunicorn==21.2.0               # WSGI HTTP Server
eventlet==0.33.3               # For SocketIO in production
gevent==23.9.1                 # Alternative async worker
